+++ bootstrap/cluster-tls/ouputs.tf
@@ -1,0 +1,3 @@
+output "cluster-nodes" {
+   value = flatten([for o in docker_container.redis-nodes : o.network_data[0].ip_address])
+}
