+++ bootstrap/cluster-tls/terraform.auto.tfvars
@@ -1,0 +1,2 @@
+
+proxy_hostname = "chanchan"
