+++ bootstrap/cluster-tls/variables.tf
@@ -1,0 +1,5 @@
+
+variable "proxy_hostname" {
+  description  = "Usually HOSTNAME of the host you are running docker on."
+  type = string
+}
