+++ bootstrap/cluster/export-cluster-vars.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+export TEST_REDIS_CLUSTER=$(terraform output -json cluster-nodes | gojq 'join(":6379,") + ":6379"' | tr -d \")
+
+unset TEST_REDIS_PRIMARY_HOST TEST_REDIS_PRIMARY_PORT TEST_REDIS_SECONDARIES TEST_REDIS_SENTINELS TEST_REDIS_SENTINEL_MASTER_NAME
