+++ redis_test.go
@@ -410,10 +410,6 @@ func checkCredsExist(t *testing.T, username, password, address string, port int)
 
 	host := address
 
-	if port < 0 {
-		host = ""
-	}
-
 	connectionDetails := map[string]interface{}{
 		"primary_host":         host,
 		"primary_port":         port,
@@ -458,10 +454,6 @@ func checkRuleAllowed(t *testing.T, username, password, address string, port int
 	t.Log("Testing checkRuleAllowed()")
 
 	host := address
-
-	if port < 0 {
-		host = ""
-	}
 
 	connectionDetails := map[string]interface{}{
 		"primary_host":         host,
@@ -652,10 +644,6 @@ func testRedisDBCreateUser_plusRole(t *testing.T, address string, port int) {
 
 	host := address
 
-	if port < 0 {
-		host = ""
-	}
-
 	connectionDetails := map[string]interface{}{
 		"primary_host":         host,
 		"primary_port":         port,
@@ -718,17 +706,13 @@ func testRedisDBCreateUser_plusRole(t *testing.T, address string, port int) {
 		t.Fatalf("Could not revoke user: %s", userResp.Username)
 	}
 }
-
+/* [TODO] groupOnly hang over from Couchbase */
 func testRedisDBCreateUser_groupOnly(t *testing.T, address string, port int) {
 	if os.Getenv("VAULT_ACC") == "" {
 		t.SkipNow()
 	}
 
 	host := address
-
-	if port < 0 {
-		host = ""
-	}
 
 	connectionDetails := map[string]interface{}{
 		"primary_host":         host,
@@ -871,7 +855,7 @@ func testRedisDBCreateUser_roleAndSelector(t *testing.T, address string, port in
 		t.Fatalf("Could not revoke user: %s", userResp.Username)
 	}
 }
-
+/* [TODO] change to test SAVE ACLFILE if supported.*/
 func testRedisDBCreateUser_persistAclFile(t *testing.T, address string, port int) {
 	if os.Getenv("VAULT_ACC") == "" {
 		t.SkipNow()
